resolve all compiler warnings
Resolve all compiler warnings when building libdispatch and
the test suite on linux.  It now compiles cleanly.

Many of the warnings were due to platform-specific sized integers
being used in printf-like contexts.  In libdispatch itself, I took
the approach of adding casts to the values to match the size/type
specified in the printf format string (under the assumption that
the expected output format should not be changed and needs to be
the same on all platforms).

+++ src/init.c
@@ -531,7 +531,7 @@ _dispatch_logv_init(void *context DISPATCH_UNUSED)
 #endif
 			dprintf(dispatch_logfile, "=== log file opened for %s[%u] at "
 					"%ld.%06u ===\n", getprogname() ?: "", getpid(),
-					tv.tv_sec, tv.tv_usec);
+					tv.tv_sec, (int)tv.tv_usec);
 		}
 	}
 }
